Move index rename outside to builder
