+++ src/vivarium_gates_nutrition_optimization_child/data/builder.py
@@ -18,6 +18,9 @@ from vivarium.framework.artifact import Artifact, EntityKey
 
 from vivarium_gates_nutrition_optimization_child.constants import data_keys
 from vivarium_gates_nutrition_optimization_child.data import loader
+from vivarium_gates_nutrition_optimization_child.data.utilities import (
+    scrub_location_level,
+)
 
 
 def open_artifact(output_path: Path, location: str) -> Artifact:
@@ -72,6 +75,8 @@ def load_and_write_data(
     else:
         logger.debug(f"Loading data for {key} for location {location}.")
         data = loader.get_data(key, location, fetch_subnationals)
+        if isinstance(data, pd.DataFrame) and ("location" in data.index.names):
+            data = scrub_location_level(data)
         if key not in artifact:
             logger.debug(f"Writing data for {key} to artifact.")
             artifact.write(key, data)
