+++ src/vivarium_gates_nutrition_optimization_child/data/loader.py
@@ -286,15 +286,7 @@ def get_data(
     else:
         subnational_ids = fetch_subnational_ids(location)
         data = mapping[lookup_key](lookup_key, subnational_ids)
-    if isinstance(data, pd.DataFrame) and ("location" in data.index.names):
-        data = scrub_location_level(data)
 
-    return data
-
-
-def scrub_location_level(data: pd.DataFrame) -> pd.DataFrame:
-    if data.index.get_level_values("location")[0] not in metadata.LOCATIONS:
-        data.index = data.index.rename({"location": "subnational"})
     return data
